Remove dashboard tests
